Move EMF parsing in a dedicated package.
