+++ jplag.frontend.emf/src/main/java/de/jplag/emf/parser/AbstractMetamodelVisitor.java
@@ -1,4 +1,4 @@
-package de.jplag.emf;
+package de.jplag.emf.parser;
 
 import java.util.List;
