+++ jplag.frontend.emf/src/main/java/de/jplag/emf/parser/TokenGeneratingMetamodelVisitor.java
@@ -1,4 +1,4 @@
-package de.jplag.emf;
+package de.jplag.emf.parser;
 
 import static de.jplag.emf.MetamodelTokenConstants.*;
 import static de.jplag.emf.MetamodelTokenConstants.ANNOTATION;
